remove some static_casts
The original reason for these is because of a clang-tidy check that
triggers on C style casts. However, {} initialization is C++. Better to
have shorter code.

+++ src/nikonmn_int.cpp
@@ -1088,9 +1088,9 @@ static void printFlashCompensationValue(std::ostream& os, const unsigned char va
     float output = 0.0;
     if (value < 128) {
       if (value != 0)
-        output = static_cast<float>(value) * static_cast<float>(-1.0);
+        output = static_cast<float>(value) * -1.0f;
     } else {
-      output = static_cast<float>(256.0) - static_cast<float>(value);
+      output = 256.0f - static_cast<float>(value);
     }
     os.precision(1);
     if (value != 0)
